requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.2.0
backports.tarfile==1.1.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cryptography==42.0.7
cycler==0.12.0
dash==2.16.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.4
docutils==0.21.2
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastjsonschema==2.19.1
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.3.3
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
h11==0.14.0
idna==3.4
image==1.5.33
importlib_metadata==7.0.2
itsdangerous==2.1.2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
keyring==25.2.0
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplot==0.1.9
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.2.0
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
mysql-connector-python==9.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
nh3==0.2.17
numpy==1.26.0
openai==0.27.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.1
pandas==2.1.1
ParallelDots==3.2.14
pigar==2.1.4
piger==0.1.1
Pillow==10.0.0
pkginfo==1.10.0
platformdirs==4.2.1
playsound==1.3.0
plotly==5.19.0
protobuf==4.25.3
pyarrow==15.0.2
PyAutoGUI==0.9.54
pycparser==2.22
pydeck==0.8.1b0
PyGetWindow==0.0.9
Pygments==2.17.2
pyloco==0.0.139
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyparsing==3.1.1
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.2
qrcode==7.4.2
readme_renderer==43.0
referencing==0.34.0
requests==2.28.2
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
selenium==4.19.0
SimpleWebSocketServer==0.1.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
streamlit==1.32.2
streamlit-drawable-canvas==0.9.3
tenacity==8.2.3
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.65.0
traitlets==5.14.3
trio==0.25.0
trio-websocket==0.11.1
twine==5.0.0
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2023.3
urllib3==1.26.15
ushlex==0.99.1
watchdog==4.0.0
websocket-client==1.8.0
Werkzeug==2.3.7
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.8.2
zipp==3.17.0